+++ src/main/java/zedly/zbot/event/InvalidBedEvent.java
@@ -2,13 +2,13 @@ package zedly.zbot.event;
 
 public class InvalidBedEvent extends Event {
 
-    private float value;
+    private double value;
 
-    public InvalidBedEvent(float value) {
+    public InvalidBedEvent(double value) {
         this.value = value;
     }
 
-    public float getValue() {
+    public double getValue() {
         return value;
     }
 }
